+++ resources/monitor.rb
@@ -1,18 +1,21 @@
-actions :install, :config, :start
+actions :create
+default_action :create
 
-property :version, String, default: '2.10.0'
-property :checksum, String, default: 'f4233783826f18606b79e5cef0686e4a9c2030146a3c7ce134f0add09f5adcb7'
-property :base_uri, String, default: 'https://github.com/prometheus/prometheus/releases/download/'
-property :home_dir, String, default: '/opt/prometheus'
-property :local_storage_path, String, default: '/var/lib/prometheus'
-property :cookbook, String, default: 'prometheus'
-
-# Config
-
-property :global, Hash, default: { "global"=>{"scrape_interval"=>"15s", "evaluation_interval"=>"15s", "external_labels"=>{"monitor"=>"Chef"}}}
-property :rule_files, Array, default: []
-
-property :scrape_configs, Array, default: [{'job_name':"node", 'file_sd_configs':[{'files':["targets.json"]}]}]
-
-property :arguments, Array, default: ['--config.file="/etc/prometheus/prometheus.yml"']
-property :template_name, String, default: "prometheus.erb"
+# Version of prometheus to install
+property :version, String, required: true
+# Checksum of the prometheus tarball archive
+property :checksum, String, required: true
+# URI to the prometheus tarball archive
+property :uri, String, required: true
+# Prometheus home directory
+property :home_dir, String, default: '/opt/prometheus', required: true
+# Cookbook name that holds prometheus.erb
+property :cookbook, String, default: 'prometheus', required: true
+# Filename of the template to be used to create /etc/default/prometheus
+property :template_name, String, default: "prometheus.erb", required: true
+# Filename of the prometheus tarball archive
+property :filename, String, required: true
+# The path where the binaries are created after extracting the tarball
+property :pathname, String, required: true
+# The arguments passed to the prometheus server command line
+property :arguments, Array, default: ['--config.file="/etc/prometheus/prometheus.yml"'], required: true