+++ test/cookbooks/test_prom/recipes/exporter_statsd.rb
@@ -1,7 +1,10 @@
-prometheus_export "statsd" do
-  action :statsd
-  checksum '9976810ae7a0e3593d6727d46d8c45a23f534e5794de816ed8309a42bb86cb34'
+prometheus_exporter "statsd_exporter" do
+  action :create
+  checksum 'c9e685db2558b96d40bb60c36420054457421fce8c6aad4b7fa14d9cff0ff04d'
   arguments ['--statsd.listen-unixgram=/tmp/statsd.sock']
-  base_uri 'https://github.com/prometheus/statsd_exporter/releases/download/v'
-  version '0.12.2'
+  uri 'https://github.com/prometheus/statsd_exporter/releases/download/v0.14.1/statsd_exporter-0.14.1.linux-amd64.tar.gz'
+  filename 'statsd_exporter-0.14.1.linux-amd64.tar.gz'
+  pathname 'statsd_exporter-0.14.1.linux-amd64'
+  binaries ['statsd_exporter']
+  start_command 'statsd_exporter'
 end